Move FETCH_TARGET_COMPUTE_URLS to common constants
