+++ app/javascript/common/constants.js
@@ -26,3 +26,11 @@ export const V2V_TARGET_PROVIDERS = [
   { name: __('Red Hat Virtualization'), id: RHV },
   { name: __('Red Hat OpenStack Platform'), id: OPENSTACK }
 ];
+
+export const FETCH_TARGET_COMPUTE_URLS = {
+  [RHV]:
+    '/api/clusters?expand=resources' +
+    '&attributes=ext_management_system.emstype,v_parent_datacenter,ext_management_system.name,hosts' +
+    '&filter[]=ext_management_system.emstype=rhevm',
+  [OPENSTACK]: '/api/cloud_tenants?expand=resources&attributes=ext_management_system.name,ext_management_system.id'
+};
