+++ app/javascript/react/screens/App/Mappings/screens/MappingWizard/components/MappingWizardClustersStep/MappingWizardClustersStepConstants.js
@@ -9,13 +9,7 @@ export const CONVERSION_HOST_TYPES = {
   [OPENSTACK]: 'ManageIQ::Providers::Openstack::CloudManager::Vm'
 };
 
-export const FETCH_TARGET_COMPUTE_URLS = {
-  [RHV]:
-    '/api/clusters?expand=resources' +
-    '&attributes=ext_management_system.emstype,v_parent_datacenter,ext_management_system.name,hosts' +
-    '&filter[]=ext_management_system.emstype=rhevm',
-  [OPENSTACK]: '/api/cloud_tenants?expand=resources&attributes=ext_management_system.name,ext_management_system.id'
-};
+export { FETCH_TARGET_COMPUTE_URLS } from '../../../../../../../../common/constants';
 
 export const QUERY_PROVIDERS_URL = '/api/providers';
